backend/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
Pillow==11.0.0
python-decouple==3.8
dj-database-url==2.2.0
whitenoise==6.8.2
gunicorn==20.1.0
